Corrections and Utils class error
This commit has all corrections except updated README and PEP-8. There
is an error due to Utils class.

+++ blog.py
@@ -17,37 +17,40 @@ jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir),
                                autoescape = True)
 
 
-secret = '111-infinity0+infinity&beyond999'
+class Utils:
+    """ contains global functions commonly used """
+    secret = '111-infinity0+infinity&beyond999'
 
-def render_str(template, **params):
-    t = jinja_env.get_template(template)
-    return t.render(params)
+    def render_str(template, **params):
+        t = jinja_env.get_template(template)
+        return t.render(params)
 
-def make_secure_val(nonsecure_val):
-    return "%s|%s" % (nonsecure_val, hmac.new(secret, nonsecure_val).hexdigest())
+    def make_secure_val(nonsecure_val):
+        return "%s|%s" % (nonsecure_val, hmac.new(secret, nonsecure_val).hexdigest())
 
-def check_secure_val(secure_val):
-    val = secure_val.split('|')[0]
-    if secure_val == make_secure_val(val):
-        return val
+    def check_secure_val(secure_val):
+        val = secure_val.split('|')[0]
+        if secure_val == make_secure_val(val):
+            return val
 
-class Handler(webapp2.RequestHandler):
+class Handler(webapp2.RequestHandler, Utils):
+    """ webapp2.RequestHandler """
     def write(self, *a, **kw):
         self.response.out.write(*a, **kw)
 
     def render_str(self, template, **params):
-        return render_str(template, **params)
+        return Utils.render_str(template, **params)
 
     def render(self, template, **kw):
         self.write(self.render_str(template, **kw))
 
     def set_secure_cookie(self, name, val):
-        cookie_val = make_secure_val(val)
+        cookie_val = Utils.make_secure_val(val)
         self.response.headers.add_header('Set-Cookie', '%s=%s; Path=/blog' % (name, cookie_val))
 
     def read_secure_cookie(self, name):
         cookie_val = self.request.cookies. get(name)
-        return cookie_val and check_secure_val(cookie_val)
+        return cookie_val and Utils.check_secure_val(cookie_val)
 
     def login(self, user):
         self.set_secure_cookie('user_id', str(user.key().id()))
@@ -62,11 +65,9 @@ class Handler(webapp2.RequestHandler):
         self.user = uid and User.by_id(int(uid))
 
 
-
 def render_post(response, post):
     response.out.write('<b>' + post.subject + '</b><br>')
     response.out.write(post.content)
-
 
 
 ### user functions
@@ -86,13 +87,17 @@ def valid_pw(name, pw, h):
 def users_key(group = 'default'):
     return db.Key.from_path('users', group)
 
+##### blog stuff
+def blog_key(name = 'default'):
+    return db.Key.from_path('blogs', name)
+
 ### user class
 class User(db.Model):
+    """ this class creates a User object in the database """
     name = db.StringProperty(required = True)
     pw_hash = db.StringProperty(required = True)
     email = db.StringProperty()
 
-    #decorator,(research this),make following functions work little different
     @classmethod
     def by_id(cls, uid):
         return cls.get_by_id(uid, parent = users_key())
@@ -110,14 +115,8 @@ class User(db.Model):
         if u and valid_pw(name, pw, u.pw_hash):
             return u
 
-
-
-##### blog stuff
-def blog_key(name = 'default'):
-    return db.Key.from_path('blogs', name)
-
-
-class Post(db.Model):
+class Post(db.Model, Utils):
+    """ this class creates a Post object related to a User in the database """
     subject = db.StringProperty(required = True)
     content = db.TextProperty(required = True)
     created = db.DateTimeProperty(auto_now_add = True)
@@ -128,10 +127,10 @@ class Post(db.Model):
 
     def render(self):
         self._render_text = self.content.replace('\n', '<br>')
-        return render_str("post.html", p= self, plks= len(self.liked_by), pdss= len(self.disd_by))
+        return Utils.render_str("post.html", p= self, plks= len(self.liked_by), pdss= len(self.disd_by))
 
-
-class Comment(db.Model):
+class Comment(db.Model, Utils):
+    """ this class creates a Comment object related to a Post and a User in the database """
     post = db.ReferenceProperty(Post, collection_name= 'post_comments')
     comcontent = db.TextProperty(required = True)
     created = db.DateTimeProperty(auto_now_add = True)
@@ -140,9 +139,10 @@ class Comment(db.Model):
 
     def render(self):
         self._render_text = self.comcontent.replace('\n', '<br>')
-        return render_str("comment.html", c= self)
+        return Utils.render_str("comment.html", c= self)
 
 class BlogFront(Handler):
+    """ this class goes to the front page with the latest 10 Posts """
     def get(self):
         posts = db.GqlQuery("select * from Post order by created desc limit 10")
         if self.user:
@@ -152,6 +152,7 @@ class BlogFront(Handler):
 
 
 class PostPage(Handler):
+    """ this class goes to a specific Post page """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         ppost = db.get(key)
@@ -199,6 +200,7 @@ class PostPage(Handler):
 
 
 class NewPost(Handler):
+    """ this class goes to the New Post Form page and handles new post submissions """
     def get(self):
         if self.user:
             self.render("newpost.html")
@@ -209,50 +211,71 @@ class NewPost(Handler):
     def post(self):
         subject = self.request.get('subject')
         content = self.request.get('content')
-
-        if subject and content:
-            p = Post(parent= blog_key(), subject= subject, content= content, username= self.user.name)
-            p.put()
-            self.redirect('/blog/%s' % str(p.key().id()))
-            return
+        if self.user:
+            if subject and content:
+                p = Post(parent= blog_key(), subject= subject, content= content, username= self.user.name)
+                p.put()
+                self.redirect('/blog/%s' % str(p.key().id()))
+                return
+            else:
+                error = "subject and content, please!"
+                self.render("newpost.html", subject=subject, content=content, error=error)
         else:
-            error = "subject and content, please!"
-            self.render("newpost.html", subject=subject, content=content, error=error)
+            self.redirect('/login')
+            return
 
 class EditPost(Handler):
+    """ this class goes to the Edit Post Form page and handles editing a post submissions """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
-        subject = post.subject
-        content = post.content
-        if self.user and self.user.name == post.username:
-            self.render("editpost.html", subject= subject, content= content)
-        elif self.user:
-            self.redirect('/blog/%s' % str(post.key().id()))
+        if post is not None:
+            subject = post.subject
+            content = post.content
+            if self.user and self.user.name == post.username:
+                self.render("editpost.html", subject= subject, content= content)
+            elif self.user:
+                self.redirect('/blog/%s' % str(post.key().id()))
+                return
+            else:
+                self.redirect('/login')
+                return
         else:
-            self.redirect('/login')
+            self.redirect('/blog')
             return
 
     def post(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         p = db.get(key)
-        subject = self.request.get('subject')
-        content = self.request.get('content')
-
-        if subject and content:
-            p.subject = subject
-            p.content = content
-            #REGION DEBUGGER used to wipe likes and dislikes
-            # p.liked_by = []
-            # p.disd_by = []
-            p.put()
-            self.redirect('/blog/%s' % str(p.key().id()))
-            return
+        if p is not None:
+            subject = self.request.get('subject')
+            content = self.request.get('content')
+            if self.user and self.user.name == p.username:
+                if subject and content:
+                    p.subject = subject
+                    p.content = content
+                    #REGION DEBUGGER used to wipe likes and dislikes
+                    # p.liked_by = []
+                    # p.disd_by = []
+                    p.put()
+                    self.redirect('/blog/%s' % str(p.key().id()))
+                    return
+                else:
+                    error = "subject and content, please!"
+                    self.render("editpost.html", subject=subject, content=content, error=error)
+            elif self.user:
+                self.redirect('/blog/%s' % str(p.key().id()))
+                return
+            else:
+                self.redirect('/login')
+                return
         else:
-            error = "subject and content, please!"
-            self.render("editpost.html", subject=subject, content=content, error=error)
+            self.redirect('/blog')
+            return
+
 
 class EditComment(Handler):
+    """ this class goes to the Edit Comment Form page and handles editing a comment submissions """
     def post(self, comment_id):
         edit_com_txtarea_name = 'comedit' + comment_id
         ctextupdate = self.request.get(edit_com_txtarea_name)
@@ -260,54 +283,72 @@ class EditComment(Handler):
         cedit = db.get(key)
         cpostid = cedit.post.key().id()
         cuename = cedit.comusername
-        if self.user and cuename == self.user.name:
-            cedit.comcontent = ctextupdate
-            cedit.put()
-            self.redirect('/blog/%s' % str(cpostid))
-            return
-        elif self.user:
-            self.redirect('/blog/%s' % str(cpostid))
-            return
+        if cedit is not None:
+            if self.user and cuename == self.user.name:
+                cedit.comcontent = ctextupdate
+                cedit.put()
+                self.redirect('/blog/%s' % str(cpostid))
+                return
+            elif self.user:
+                self.redirect('/blog/%s' % str(cpostid))
+                return
+            else:
+                self.redirect('/login')
+                return
         else:
-            self.redirect('/login')
+            self.redirect('/blog/%s' % str(cpostid))
             return
 
 class DeletePost(Handler):
+    """ this class handles deleting a post submission """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         pdelete = db.get(key)
         pudname = pdelete.username
-        if self.user and pudname == self.user.name:
-            pdelete.delete()
-            for c in pdelete.post_comments:
-                c.delete()
-            self.redirect('/blog')
-            return
-        elif self.user:
-            self.redirect('/blog/%s' % str(pdelete.key().id()))
-            return
+        if pdelete is not None:
+            if self.user and pudname == self.user.name:
+                pdelete.delete()
+                for c in pdelete.post_comments:
+                    c.delete()
+                self.redirect('/blog')
+                return
+            elif self.user:
+                self.redirect('/blog/%s' % str(pdelete.key().id()))
+                return
+            else:
+                self.redirect('/login')
+                return
         else:
-            self.redirect('/login')
+            self.redirect('/blog/%s' % str(pdelete.key().id()))
             return
 
 class DeleteComment(Handler):
+    """ this class handles deleting a comment submission """
     def get(self, comment_id):
         key = db.Key.from_path('Comment', int(comment_id))
         cdelete = db.get(key)
         cpostid = cdelete.post.key().id()
         cudname = cdelete.comusername
-        if self.user and cudname == self.user.name:
-            cdelete.delete()
-            self.redirect('/blog/%s' % str(cpostid))
-            return
-        elif self.user:
+        if cdelete is not None and cpostid is not None:
+            if self.user and cudname == self.user.name:
+                cdelete.delete()
+                self.redirect('/blog/%s' % str(cpostid))
+                return
+            elif self.user:
+                self.redirect('/blog/%s' % str(cpostid))
+                return
+            else:
+                self.redirect('/login')
+                return
+        elif cdelete is not None and cpostid == None:
             self.redirect('/blog/%s' % str(cpostid))
             return
         else:
-            self.redirect('/login')
+            self.redirect('/blog')
             return
 
 class LikePost(Handler):
+    """ this class handles Liking a post submission """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         plike = db.get(key)
@@ -328,6 +369,7 @@ class LikePost(Handler):
             return
 
 class DisPost(Handler):
+    """ this class handles Disliking a post submission """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         pdis = db.get(key)
@@ -348,19 +390,20 @@ class DisPost(Handler):
             return
 
 
-USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
 def valid_username(username):
+    USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
     return username and USER_RE.match(username)
 
-PASS_RE = re.compile(r"^.{3,20}$")
 def valid_password(password):
+    PASS_RE = re.compile(r"^.{3,20}$")
     return password and PASS_RE.match(password)
 
-EMAIL_RE  = re.compile(r'^[\S]+@[\S]+\.[\S]+$')
 def valid_email(email):
+    EMAIL_RE  = re.compile(r'^[\S]+@[\S]+\.[\S]+$')
     return not email or EMAIL_RE.match(email)
 
 class Signup(Handler):
+    """ handles New User signup submission """
     def get(self):
         self.render("signup-form.html")
 
@@ -399,6 +442,7 @@ class Signup(Handler):
 
 
 class Register(Signup):
+    """ confirms Signup submission does not contain an existing username in the database """
     def done(self):
         #check that user doesn't exist
         u = User.by_name(self.username)
@@ -414,6 +458,7 @@ class Register(Signup):
 
 
 class Login(Handler):
+    """ handles User Login submission """
     def get(self):
         #auto-logout a loggedin user if they wander to login page manually
         self.logout()
@@ -434,6 +479,7 @@ class Login(Handler):
 
 
 class Logout(Handler):
+    """ handles user logout safely """
     def get(self):
         self.logout()
         self.redirect('/login')
@@ -441,6 +487,7 @@ class Logout(Handler):
 
 
 class Welcome(Handler):
+    """ goes to the user welcom page """
     def get(self):
         #the following 'user' is from the initialize function above
         if self.user:
@@ -450,6 +497,7 @@ class Welcome(Handler):
             return
 
 class MainPage(Handler):
+    """ goes to the signup page """
     def get(self):
         self.redirect('/signup')
